pyproject.toml
[project]
name = "python-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.8",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "python-dotenv>=1.1.1",
    "sql>=2022.4.0",
    "uvicorn>=0.37.0",
]
